pyproject.toml
[tool.poetry]
name = "intics"
version = "2.0.0"
description = ""
authors = ["SOSP <sosp@halo.com>"]

[tool.poetry.dependencies]
python = "=3.8.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
PyPDF2 = "^1.26.0"
opencv-python = "^4.5.5"
pdf2image = "^1.16.0"
pytesseract = "^0.3.9"
scipy = "^1.8.1"
pandas= "^1.2.3"
imutils= "^0.5.4"
matplotlib = "^3.5.2"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.3"
colorlog = "^6.6.0"
torch = "^1.12.0"
torchvision = "^0.13.0"
tqdm = "^4.64.0"
albumentations = "^1.2.1"
efficientnet-pytorch = "^0.7.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
